Inline or removing remaining CollectionUtil methods
